requirements.txt
chainlit
openai
azure-search-documents
python-dotenv
nest_asyncio